+++ lib/core/di/di.dart
@@ -1,4 +1,9 @@
 import 'package:get_it/get_it.dart';
+import 'package:logger/logger.dart';
+
+import '../utils/logger/app_logger.dart';
+import '../utils/logger/app_logger_impl.dart';
+import '../utils/logger/logger_setup.dart';
 
 /// Global instance of the GetIt service locator for dependency management.
 final di = GetIt.instance;
@@ -7,4 +12,10 @@ final di = GetIt.instance;
 void setupDI() {
   // Register your dependencies here, for example:
   // di.registerLazySingleton<SomeService>(() => SomeServiceImpl());
+
+  // Logger
+  final logger = createLogger();
+
+  di.registerLazySingleton<Logger>(() => logger);
+  di.registerLazySingleton<AppLogger>(() => AppLoggerImpl(di<Logger>()));
 }